refactoring: data 폴더 삭제
